# public/api/v1/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NeuralNexus API",
        "description": "API for NeuralNexus site and services",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.neuralnexus.dev/api/v1",
            "description": "Production server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "summary": "Login to the API",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthSessionResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/401Unauthorized"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "summary": "Logout of the API",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthSessionResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/401Unauthorized"
                    }
                }
            }
        },
        "/bee-name-generator/name": {
            "get": {
                "summary": "Get a random bee name",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BNGNameResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/500InternalServerError"
                    }
                }
            }
        },
        "/bee-name-generator/name/{name}": {
            "post": {
                "summary": "Get a specific bee name",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BNGNameSlug"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BNGNameResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/500InternalServerError"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific bee name",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BNGNameSlug"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BNGNameResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/500InternalServerError"
                    }
                }
            }
        },
        "/bee-name-generator/suggestion/{param}": {
            "get": {
                "summary": "Get bee name suggestions",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bee name suggestions response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/401Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/500InternalServerError"
                    }
                }
            },
            "post": {
                "summary": "Suggest a bee name",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BNGNameSlug"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BNGNameResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/500InternalServerError"
                    }
                }
            },
            "put": {
                "summary": "Accept a bee name suggestion",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BNGNameSlug"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BNGNameResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/401Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/500InternalServerError"
                    }
                }
            },
            "delete": {
                "summary": "Reject a bee name suggestion",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BNGNameSlug"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/BNGNameResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/400BadRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/401Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/500InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Problem": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "instance": {
                        "type": "string"
                    }
                }
            },
            "BNGName": {
                "type": "object",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    }
                }
            }
        },
        "parameters": {
            "BNGNameSlug": {
                "name": "name",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token"
            }
        },
...
        "responses": {
            "400BadRequest": {
                "description": "Bad request",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    },
                    "application/problem+xml": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
            "401Unauthorized": {
                "description": "Unauthorized",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    },
                    "application/problem+xml": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
            "500InternalServerError": {
                "description": "Internal server error",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    },
                    "application/problem+xml": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
            "AuthSessionResponse": {
                "description": "Session response",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "session_id": {
                                    "type": "string"
                                },
                                "user_id": {
                                    "type": "string"
                                },
                                "permissions": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "iat": {
                                    "type": "integer"
                                },
                                "lua": {
                                    "type": "integer"
                                },
                                "exp": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "application/xml": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "session_id": {
                                    "type": "string"
                                },
                                "user_id": {
                                    "type": "string"
                                },
                                "permissions": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "iat": {
                                    "type": "integer"
                                },
                                "lua": {
                                    "type": "integer"
                                },
                                "exp": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "BNGNameResponse": {
                "description": "Bee name response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BNGName"
                        }
                    },
                    "application/xml": {
                        "schema": {
                            "$ref": "#/components/schemas/BNGName"
                        }
                    }
                }
            }
        }
    }
}